Fix examples and snapshot code of MemStorageCore (#409)
Close #408.

The commit index of LightReady is not handled in examples.
But the test can not pass is mainly due to the bug of generating snapshot in MemStorageCore.
The term in snapshot meta should not be the hard_state.term. It should be the log term of meta.index.

By the way, the snapshot has an assumption that all entries whose index is less than hard_state.commit has been applied, which is not true when using async ready. So I leave a TODO and I will resolve it soon.

Signed-off-by: gengliqi <gengliqiii@gmail.com>